modules/euclid/requirements.txt
backports-abc==0.4
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.5.0.1
certifi==2016.2.28
configparser==3.3.0.post2
cycler==0.10.0
dask==0.8.1
decorator==4.0.9
Django==1.9.6
entrypoints==0.2.1
euclid==0.0.1
functools32==3.2.3.post2
image==1.5.3
imagemat==0.0.1
ipykernel==4.3.1
ipython==4.2.0
ipython-genutils==0.1.0
ipywidgets==5.1.4
Jinja2==2.8
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.2.2
jupyter-console==4.1.1
jupyter-core==4.1.0
MarkupSafe==0.23
matplotlib==1.5.1
mistune==0.7.2
Mosek==7.1.49
nbconvert==4.2.0
nbformat==4.0.1
notebook==4.2.0
numpy==1.11.0
pathlib2==2.1.0
pexpect==4.0.1
pickleshare==0.7.2
Pillow==3.2.0
ptyprocess==0.5.1
pycairo==1.10.0
Pygments==2.1.3
pyparsing==2.1.1
python-dateutil==2.5.3
pytz==2016.4
pyzmq==15.2.0
qtconsole==4.2.1
roibuddy==1.0.0
scipy==0.17.1
sima==1.3.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
spread==0.0.1
terminado==0.6
tornado==4.3
tqdm==4.7.0
traitlets==4.2.1
widgetsnbextension==1.2.2
